PeripheralCecAdapter: remove HDMI port and use physical address setting
By the HDMI port option it was not possible to represent all hardware setups available on user side.
Also the HDMI port setting was transfered to the corresponding physical address by libCEC on the first init anyway.

The physical address in Kodi/input is now set only by user, not by auto detection anymore.
If the user change the physical address it will be forced to use user setting instead auto detection.